Added spec for Primitives serde + fixed a state issue with test

+++ suuchi-core/src/main/scala/in/ashwanthkumar/suuchi/store/PrimitivesSerDeUtils.scala
@@ -8,6 +8,6 @@ object PrimitivesSerDeUtils {
   * */
   def longToBytes(instance: Long) = ByteBuffer.allocate(8).putLong(instance).array()
   def intToBytes(instance: Int) = ByteBuffer.allocate(4).putInt(instance).array()
-  def bytesToInt(bytes: Array[Byte]) = ByteBuffer.wrap(bytes).getLong
+  def bytesToInt(bytes: Array[Byte]) = ByteBuffer.wrap(bytes).getInt
   def bytesToLong(bytes: Array[Byte]) = ByteBuffer.wrap(bytes).getLong
 }
